Add nested landing page for each event
